Move handle_event_test to components module
The handle_event_test utility function was only used in the components
module, so move it to that module and remove the dependency on module.
