+++ src/components/mod.rs
@@ -2,3 +2,6 @@ pub mod choice;
 pub mod confirm;
 pub mod edit;
 pub mod help;
+
+#[cfg(test)]
+mod testutil;
