+++ src/module/testutil/mod.rs
@@ -1,4 +1,3 @@
 mod assert_process_result;
-mod handle_event_test;
 
-pub use self::{assert_process_result::_assert_process_result, handle_event_test::handle_event_test};
+pub use self::assert_process_result::_assert_process_result;
